refactor: Refactor authorisation forms layout
