+++ src/index.css
@@ -32,12 +32,6 @@
   padding: 0 20px;
 }
 
-.center-children {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
 .column {
   box-sizing: border-box;
   width: 600px;
