+++ src/pages/page-layout.module.css
@@ -1,8 +1,15 @@
-.container {
-  flex-direction: column;
+.formContainer {
   margin-right: auto;
   margin-left: auto;
   width: 480px;
+  text-align: center;
+}
+
+.form {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
 }
 
 .inputWrapper {
