--- src/pages/register.module.css
@@ -1,25 +1,0 @@
-.container {
-  flex-direction: column;
-  margin-right: auto;
-  margin-left: auto;
-  width: 480px;
-}
-
-.inputWrapper {
-  display: flex;
-  width: 100%;
-  justify-content: center;
-}
-
-.link {
-  color: var(--colors-interface-accent);
-  text-decoration: none;
-}
-
-.link:hover {
-  color: var(--colors-interface-success);
-}
-
-.link:active {
-  opacity: 0.5;
-}
